libcontainer: rootfs: split setupRootfs
In order to allow for late console setup, split setupRootfs into the
"preparation" section where all of the mounts are created and the
"finalize" section where we pivot_root and set things as ro. In between
the two we can set up all of the console mountpoints and symlinks we
need.

Signed-off-by: Aleksa Sarai <asarai@suse.de>