nsenter: cloned_binary: smoke-test memfd_create before using it
On some RHEL kernels, memfd_create(2) has a busted port which results in
seals not working properly. Rather than failing outright, we can try to
detect this by smoke-testing the sealing functionality of
memfd_create(2).

It should be noted this isn't fool-proof (it's possible that there is a
porting bug related to exec-ing a memfd) but should be good enough as a
simple and cheap sanity-check.

Signed-off-by: Aleksa Sarai <asarai@suse.de>