cgroups: systemd: make use of Device*= properties
It seems we missed that systemd added support for the devices cgroup, as
a result systemd would actually *write an allow-all rule each time you
did 'runc update'* if you used the systemd cgroup driver. This is
obviously ... bad.

So we simply generate the cgroupv1-style rules (which is what systemd's
DevicesAllow wants) and default to a deny-all ruleset. Unfortunately
it turns out that systemd is susceptible to the same spurrious error
failure that we were, so that problem is out of our hands for systemd
cgroup users.

However, systemd has a similar bug to the one fixed in [1]. It will
happily write a disruptive deny-all rule when it is not necessary.
Unfortunately, we cannot even use devices.Emulator to generate a minimal
set of transition rules because the DBus API is limited (you can only
clear or append to the DevicesAllow= list -- so we are forced to always
clear it). To work around this, we simply freeze the container during
SetUnitProperties.

[1]: b70aacc28f43 ("cgroupv1: devices: use minimal transition rules with devices.Emulator")

Signed-off-by: Aleksa Sarai <asarai@suse.de>